# draft: new field input format