--- a/log_test.go
+++ b/log_test.go
@@ -58,11 +58,11 @@ func TestLogFormat(t *testing.T) {
 	buf.Reset()
 
 	// Log with field.
-	fields := Fields{
-		"stack": "testing",
+	fields := F{
+		"stack", "testing",
 	}
 	fl := l.WithFields(fields)
-	assert.Equal(t, fl.fields, fields)
+	assert.Equal(t, fl.fields, []F{fields})
 	fl.Warn("testing fields")
 	assert.Contains(t, buf.String(), `level=warn message="testing fields" stack=testing`, "warning log")
 	buf.Reset()
@@ -71,7 +71,7 @@ func TestLogFormat(t *testing.T) {
 	fakeErr := errors.New("this is a fake error")
 	el := l.WithError(fakeErr)
 	// Check if error key exists.
-	assert.Equal(t, el.fields["error"], fakeErr.Error())
+	assert.Equal(t, el.fields[0].V, fakeErr.Error())
 	el.Error("testing error")
 	assert.Contains(t, buf.String(), `level=error message="testing error" error="this is a fake error"`, "error log")
 	buf.Reset()
@@ -96,6 +96,6 @@ func TestLoggerConcurrency(t *testing.T) {
 
 func genLogs(l *Logger) {
 	for i := 0; i < 100; i++ {
-		l.WithFields(Fields{"index": strconv.FormatInt(int64(i), 10)}).Info("random log")
+		l.WithFields(F{"index", strconv.FormatInt(int64(i), 10)}).Info("random log")
 	}
 }