Benchstat:

```
name                      old time/op    new time/op    delta
NoField-8                    179ns ± 3%     182ns ± 7%     ~     (p=0.687 n=3+3)
OneField-8                   283ns ± 1%     214ns ± 5%  -24.39%  (p=0.008 n=3+3)
ThreeFields-8                338ns ± 2%     268ns ± 1%  -20.82%  (p=0.002 n=3+3)
ErrorField-8                 381ns ±13%     270ns ±11%  -29.14%  (p=0.029 n=3+3)
HugePayload-8               1.11µs ± 6%    0.69µs ± 2%  -37.92%  (p=0.004 n=3+3)
ThreeFields_WithCaller-8     932ns ± 4%     860ns ±13%     ~     (p=0.337 n=3+3)
NoField_WithColor-8          320ns ± 1%     298ns ±10%     ~     (p=0.316 n=3+3)
OneField_WithColor-8         441ns ± 9%     351ns ± 7%  -20.39%  (p=0.026 n=3+3)
ThreeFields_WithColor-8      569ns ±19%     419ns ± 1%     ~     (p=0.112 n=3+3)
ErrorField_WithColor-8       500ns ±10%     373ns ± 2%  -25.36%  (p=0.038 n=3+3)
HugePayload_WithColor-8     1.24µs ± 7%    0.88µs ± 8%  -28.93%  (p=0.003 n=3+3)

name                      old alloc/op   new alloc/op   delta
NoField-8                    0.00B          0.00B          ~     (zero variance)
OneField-8                    336B ± 0%       32B ± 0%     ~     (zero variance)
ThreeFields-8                 336B ± 0%       96B ± 0%     ~     (zero variance)
ErrorField-8                  368B ± 0%       80B ± 0%     ~     (zero variance)
HugePayload-8                 919B ± 0%      320B ± 0%     ~     (zero variance)
ThreeFields_WithCaller-8      656B ± 0%      416B ± 0%     ~     (zero variance)
NoField_WithColor-8          0.00B          0.00B          ~     (zero variance)
OneField_WithColor-8          336B ± 0%       32B ± 0%     ~     (zero variance)
ThreeFields_WithColor-8       336B ± 0%       96B ± 0%     ~     (zero variance)
ErrorField_WithColor-8        368B ± 0%       80B ± 0%     ~     (zero variance)
HugePayload_WithColor-8       919B ± 0%      320B ± 0%     ~     (zero variance)

name                      old allocs/op  new allocs/op  delta
NoField-8                     0.00           0.00          ~     (zero variance)
OneField-8                    2.00 ± 0%      1.00 ± 0%     ~     (zero variance)
ThreeFields-8                 2.00 ± 0%      1.00 ± 0%     ~     (zero variance)
ErrorField-8                  4.00 ± 0%      3.00 ± 0%     ~     (zero variance)
HugePayload-8                 3.00 ± 0%      1.00 ± 0%     ~     (zero variance)
ThreeFields_WithCaller-8      7.00 ± 0%      6.00 ± 0%     ~     (zero variance)
NoField_WithColor-8           0.00           0.00          ~     (zero variance)
OneField_WithColor-8          2.00 ± 0%      1.00 ± 0%     ~     (zero variance)
ThreeFields_WithColor-8       2.00 ± 0%      1.00 ± 0%     ~     (zero variance)
ErrorField_WithColor-8        4.00 ± 0%      3.00 ± 0%     ~     (zero variance)
HugePayload_WithColor-8       3.00 ± 0%      1.00 ± 0%     ~     (zero variance)
```