@rhnvrm 2022-07-06
Closing in favor of `func(...any)` kind of exposed API.